+++ task/orchestrator.go
@@ -4,20 +4,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
-	"os/exec"
-	"os/user"
-	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
-	"syscall"
 	"time"
 
 	"github.com/circleci/ex/o11y"
 	"github.com/hashicorp/go-reap"
 
 	"github.com/circleci/runner-init/clients/runner"
+	"github.com/circleci/runner-init/task/cmd"
 )
 
 type Orchestrator struct {
@@ -26,9 +21,9 @@ type Orchestrator struct {
 	gracePeriod  time.Duration
 
 	ready      atomic.Bool
-	agentPid   atomic.Int64
+	entrypoint cmd.Command
+	taskAgent  cmd.Command
 	cancelTask context.CancelFunc
-	cancelMu   sync.RWMutex
 	reapMu     sync.RWMutex
 }
 
@@ -41,7 +36,6 @@ func NewOrchestrator(config Config, runnerClient *runner.Client, gracePeriod tim
 		config:       config,
 		runnerClient: runnerClient,
 		gracePeriod:  gracePeriod,
-		cancelTask:   func() {},
 	}
 }
 
@@ -53,16 +47,14 @@ func (o *Orchestrator) Run(parentCtx context.Context) (err error) {
 	go o.reapChildProcesses(ctx)
 
 	defer func() {
-		cleanupErr := o.cleanup(ctx)
-		if cleanupErr != nil {
-			cleanupErr = fmt.Errorf("error on shutdown: %w", cleanupErr)
-		}
-		err = o.handleRunErrors(ctx, errors.Join(cleanupErr, err))
+		err = o.shutdown(ctx, err)
 	}()
 
 	if len(o.config.Cmd) > 0 {
-		// If a custom command is specified, execute it in the background
-		go o.executeCmd(ctx)
+		// If a custom entrypoint is specified, execute it in the background
+		if err := o.executeEntrypoint(ctx); err != nil {
+			return err
+		}
 	}
 
 	// Signal the orchestrator is ready and will start the task agent process
@@ -83,7 +75,7 @@ func (o *Orchestrator) Run(parentCtx context.Context) (err error) {
 	case err := <-errCh:
 		return err
 	case <-parentCtx.Done():
-		// If the parent context is canceled, wait for the termination grace period before shutting down.
+		// If the parent context is cancelled, wait for the termination grace period before shutting down.
 		// This is in case the task completes within that period.
 		select {
 		case err := <-errCh:
@@ -96,9 +88,6 @@ func (o *Orchestrator) Run(parentCtx context.Context) (err error) {
 }
 
 func (o *Orchestrator) taskContext(ctx context.Context) context.Context {
-	o.cancelMu.Lock()
-	defer o.cancelMu.Unlock()
-
 	// Copy the O11y provider to a new context that can be separately cancelled.
 	// This ensures we can drain the task on shutdown of the agent even if the parent context was cancelled,
 	// but still make sure any task resources are released.
@@ -106,125 +95,55 @@ func (o *Orchestrator) taskContext(ctx context.Context) context.Context {
 	return ctx
 }
 
-func (o *Orchestrator) executeCmd(ctx context.Context) {
-	cmd := o.config.Cmd
-	c := o.newCmd(ctx, cmd)
-	if err := c.Run(); err != nil {
-		o11y.LogError(ctx, "error running custom command", err, o11y.Field("cmd", cmd))
+func (o *Orchestrator) executeEntrypoint(ctx context.Context) error {
+	c := o.config.Cmd
+	o.entrypoint = cmd.New(ctx, c, "")
+
+	if err := o.entrypoint.Start(); err != nil {
+		return fmt.Errorf("error starting custom entrypoint %s: %w", c, err)
 	}
+	return nil
 }
 
 func (o *Orchestrator) executeAgent(ctx context.Context) error {
-	agent := o.config.Agent()
-	c := o.newCmd(ctx, agent.Cmd, agent.Env...)
+	cfg := o.config
+	agent := cfg.Agent()
 
-	if err := o.loadToken(c); err != nil {
-		return retryableErrorf("failed to load task token: %w", err)
-	}
+	o.taskAgent = cmd.New(ctx, agent.Cmd, cfg.User, agent.Env...)
 
-	// Start and wait for the task agent process to exit
-	if err := c.Start(); err != nil {
+	if err := o.taskAgent.StartWithStdin([]byte(cfg.Token.Raw())); err != nil {
 		return retryableErrorf("failed to start task agent command: %w", err)
 	}
-	if c.Process == nil {
-		return retryableErrorf("no process associated with task agent command")
-	}
-	// Store the task agent PID so that we can inspect the process later on cleanup
-	o.agentPid.Store(int64(c.Process.Pid))
-	if err := c.Wait(); err != nil {
+
+	if err := o.taskAgent.Wait(); err != nil {
 		return fmt.Errorf("task agent command exited with an unexpected error: %w", err)
 	}
 
 	return nil
 }
 
-func (o *Orchestrator) loadToken(c *exec.Cmd) error {
-	// Pass the task token to the task agent process through its stdin pipe
-	w, err := c.StdinPipe()
-	if err != nil {
-		return fmt.Errorf("unexpected error on stdin pipe for task agent command: %w", err)
-	}
+func (o *Orchestrator) shutdown(ctx context.Context, runErr error) (err error) {
 	defer func() {
-		_ = w.Close()
-	}()
-
-	_, err = w.Write([]byte(o.config.Token.Raw()))
-	if err != nil {
-		return fmt.Errorf("failed to write task token to stdin pipe for task agent command: %w", err)
-	}
-
-	return nil
-}
-
-func (o *Orchestrator) newCmd(ctx context.Context, cmd []string, env ...string) *exec.Cmd {
-	//#nosec:G204 // this is intentionally setting up a command
-	c := exec.CommandContext(ctx, cmd[0], cmd[1:]...)
-
-	for _, env := range os.Environ() {
-		if strings.HasPrefix(env, "CIRCLECI_GOAT") {
-			// Prevent orchestrator configuration from being injected in the task environment
-			continue
+		if err != nil {
+			err = fmt.Errorf("error on shutdown: %w", err)
 		}
-		c.Env = append(c.Env, env)
-	}
-	if env != nil {
-		c.Env = append(c.Env, env...)
-	}
+		err = errors.Join(err, runErr)
+		if err != nil {
+			err = o.handleErrors(ctx, err)
+		}
 
-	c.Stdout = os.Stdout
-	c.Stderr = os.Stderr
-
-	o.maybeSwitchUser(ctx, c)
-
-	return c
-}
-
-func (o *Orchestrator) maybeSwitchUser(ctx context.Context, c *exec.Cmd) {
-	username := o.config.User
-	if username == "" {
-		return
-	}
-
-	usr, err := user.Lookup(username)
-	if err == nil {
-		uid, _ := strconv.Atoi(usr.Uid)
-		gid, _ := strconv.Atoi(usr.Gid)
-		c.SysProcAttr = &syscall.SysProcAttr{}
-		c.SysProcAttr.Credential = &syscall.Credential{Uid: uint32(uid), Gid: uint32(gid)}
-	} else {
-		o11y.LogError(ctx, "failed to lookup user", err, o11y.Field("username", username))
-	}
-}
-
-func (o *Orchestrator) cleanup(_ context.Context) error {
-	defer func() {
-		// Cancelling the context terminates the task agent and custom entrypoint commands (if still running)
-		o.cancelMu.RLock()
-		defer o.cancelMu.RUnlock()
 		o.cancelTask()
 	}()
 
-	pid := o.agentPid.Load()
-	if pid > 0 {
-		if p, err := os.FindProcess(int(pid)); err == nil {
-			if err := p.Signal(os.Signal(syscall.Signal(0))); err == nil {
-				return fmt.Errorf("task agent process is still running, which could interrupt the task. " +
-					"Possible reasons include the Pod being evicted or deleted")
-
-			} else if !errors.Is(err, os.ErrProcessDone) {
-				return fmt.Errorf("unexpected error while signaling task agent process: %w", err)
-			}
-		}
+	isRunning, err := o.taskAgent.IsRunning()
+	if isRunning {
+		return fmt.Errorf("task agent process is still running, which could interrupt the task. " +
+			"Possible reasons include the Pod being evicted or deleted")
 	}
-
-	return nil
+	return err
 }
 
-func (o *Orchestrator) handleRunErrors(ctx context.Context, err error) error {
-	if err == nil {
-		return nil
-	}
-
+func (o *Orchestrator) handleErrors(ctx context.Context, err error) error {
 	ctx = o11y.WithProvider(context.Background(), o11y.FromContext(ctx))
 	c := o.config
 
@@ -249,7 +168,7 @@ func (o *Orchestrator) handleRunErrors(ctx context.Context, err error) error {
 	return errors.Join(failErr, unclaimErr, err)
 }
 
-var reapTime = 10 * time.Second // can be overridden by tests
+var reapTime = 30 * time.Second // can be overridden by tests
 
 // Reap any child processes that may be spawned by the task
 func (o *Orchestrator) reapChildProcesses(ctx context.Context) {